backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.10
pydantic-settings==2.5.2
pydantic[email]==2.9.2
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
apscheduler==3.10.4
pywebpush==2.0.0
anthropic==0.40.0
python-multipart==0.0.12
httpx==0.27.2
